refactor(util): consolidate duplicate hasUnresolvedPlaceholders function
Extract hasUnresolvedPlaceholders into shared pkg/util package to
eliminate code duplication between pkg/templates/execute.go and
pkg/owasp/mutation.go.
